Added the Interview Constructor
After playing around with Docassemble, it seemed like there were some things about interviews that could potentially be abstracted. Basically, I figured that there are a few types of questions that exist in interviews:
• Basic Q: A basic, top-level question
• Sub-Q: If the user selects “I don’t know” on a Q, 1+ sub-Qs will take them through determining the answer to their super-Q
• Detail Q: A special type of sub-Q, that isn’t supposed to help determine the answer to its super-Q; instead, it is just to add some detail/improve the user’s case

So—I wrote this “interview constructor”, to try to capture some of the object-oriented aspects of creating legal interviews.